// src/types/logTypes.ts
export type LogType = {
  id: string
  orderId: string
  orderDate: string
  deliveryCompany: string
  url: string
  errorTime: string
  errorMessage: string
  method: string
  webhookText: string
  nextAttemptTime: string
  numberOfAttempts: number
  pickupId: string
  createdDate: string
  shipmentNumber: string
  status: string
  logStatus: string
}